Funciones

Index

NowcastingHuascaran.generate_climarray

NowcastingHuascaran.generate_dataset

NowcastingHuascaran.ncwrite_compressed

NowcastingHuascaran.read_fed

Dataset

NowcastingHuascaran.generate_dataset — Function

generate_dataset(climarr::AbstractArray, dimensions::NTuple{3}; 
                 radius::AbstractFloat, 
                 min_neighbors::Int, 
                 min_cluster_size::Int, 
                 t_scale::AbstractFloat, 
                 windows::NTuple{3}=(0,0,0), 
                 threshold::Int=0, 
                 padding::NTuple{3}=(0,0,0))

Crea un diccionario con los elementos componentes del dataset.

Los argumentos radius, min_neighbors y min_cluster_size son pasados directamente a Clustering.dbscan.

Argumentos

climarr : ClimArray que contiene la data con dimensiones longitud, latitud y tiempo

dimensions : tupla que representa las dimensiones de los parches de salida

t_scale::AbstractFloat : factor para multiplicar la dimensión temporal antes de clusterizar

windows::NTuple{3} = (0, 0, 0) : dimensiones de las ventanas moviles que se aplicarán

threshold::Int = 0 : umbral para la binarización

padding::NTuple{3} = (0, 0, 0) : tamaño del padding en cada dimensión

Detalles

El argumento climarr debe ser un ClimArray con 3 dimensiones ordenadas: longitud, latitud y tiempo.

Se aplica clusterización al array climarr, usando dbscan y tratando a la dimensión T como una tercera dimensión espacial escalada con el factor t_scale. Se calculan bounding boxes de tamaño dimensions para cada cluster obtenido.

Para cada dimensión si el tamaño del cluster es mayor al del bounding box propuesto aplicará una ventana movil en esa dimensión. La cantidad de elementos a desplazarse en la dimensión está indicado por windows, si este valor es 0, se tomará el límite inferior del cluster. En caso contrario se aplicará padding en la dimensión y se obtendrán nuevas bounding boxes de tamaño dimensions.

La función generate_dataset devolverá un diccionario con 4 entradas:

FED : un arreglo de 5 dimensiones (Lon×Lat×1×N×T) donde la 3ra dimensión está vacía por conveniencia

lon : una matriz (Lon×N) que almacena la longitud de cada elemento en la grilla

lat : una matriz (Lat×N) que almacena la latitud de cada elemento en la grilla

time : una matriz (T×N) que almacena el tiempo en segundos desde el unix epoch como Float64 de cada elemento en la grilla

NowcastingHuascaran.read_fed — Function

read_fed(file::AbstractString)

Regenera el ClimArray producido por generate_climarray almacenado en formato HDF5.

NowcastingHuascaran.generate_climarray — Function

generate_climarray(records::AbstractVector{FlashRecords},
                   s::Quantity, 
                   t::Period;
                   N=PERU_N,
                   S=PERU_S,
                   E=PERU=E,
                   W=PERU_W)

Crea un ClimArray a partir de los FlashRecords recibidos.

Argumentos

records : lista de FlashRecords a agrupar

s::Quantity : resolución espacial del arreglo resultante

t::Period : resolución temporal del arreglo resultante

N, S, E, W : bordes para generar la grilla de acumulación de Flashes

Detalle

Se calcula el tamaño del arreglo resultante. Una grilla 3D rectangular Float32 es creada con esquinas definidas por N, S, E, W (por defecto los bordes del mapa peruano). Se acumula la ocurrencia de Flashes en la grilla, finalmente es transformada en un ClimArray con nombre "FED".

NowcastingHuascaran.ncwrite_compressed — Function

ncwrite_compressed(file::String, Xs; globalattr = Dict(), deflatelevel::Int=1)

Adaptado de ClimateBase.ncwrite y permite especificar un nivel de compresión.